Squashed all the migrations for a nice speedup
Doing all the migrations have been really slow. This has been a problem
while writing code and testing at the same time, because while running
the tests we have to do the migrations each time.

This commit squashes and tidies the migrations so that they are much
faster. On my machine it gives a speedup of 3-4 seconds.