Switch to using django-wiki from pypi and upgrade to newer version (0.3b3)
People were having problems editing pages on the django-wiki.
I was because of some template error somewhere.

I suspect it had to do with our upgrade to django 1.11, because
it says on django-wikis github page that django-wiki 0.2 don't support
django 1.11.

Fixing this by upgrading to beta version of 0.3.
In the process I have reverted to using django-wiki from pypi instead
of our fork of it.

I have never seen the need for having our own fork of django-wiki.
If we need the extra plugins that were added to our fork then we should
put it in a different python package and git repository.