Create nablapps.core and make nablaweb/ not be a django app.
nablapps.core is an app for all that does not fit in any other app.

This commit creates nablapps.core and moves all
django-appish things from nablaweb/ to nablapps.core.

The reason for this change is that similar things should
lie in the same place and different things should lie in
different places.
A django app and django project configuration are different things.

It is also nice that all apps in nablaweb are located in the same folder,
that is the nablapps folder.