Add base template for advent 2018
The template is a copy of the template from 2016, but the image
associated with a door is used as the icon for that door, instead of the
frame. If there is no image, the cycle of images used in the 2016
template is used as a fallback.

Not sure if this is the best way to do it, as it creates unessecessary
redundancy, but I don't know any better.