Refactor waiting lists for events
EventRegistrations don't use the field 'number' anymore to sort the waiting list.
The waiting list is only sorted by the id, i.e. the time
the registrations was created.

This commit also removes the other manager for EventRegistrations,
mostly because there are no longer any complicated code to assure that
the waiting list is ordered correctly.