Implement session-based solution for quizreplies
Quizreplies are no longer created when the user opens the quiz, which
alleviates the issue of "empty" responses.
When the user starts a quiz a uuid is generated for that quizreply, and
the start-time is stored in the user's session. The uuid for the
quizreply is included as a hidden value in the POST request, and is used
to recall the start-time when the user submits their answers.