Hides form for creating thread in common channels.
Common channels are meant to be more of a one way news feed from e.g.
the board, and new threads here are currently planned to be made in
admin. Just hiding the form in the template is not a very secure
sollution, but It can be improved on eventually.