Alters "fag/subject" field of exchange objects.
Now instead of a simple CharField for specifying subjects,
each subject is it's own instance of model Subject, with a
foreign key relation to the exhange object. This allows for better a
better experience in the admin panel, and also makes it easier to
display the subjects in a readable, consistent manner.